Upstream version 0.10.0~rc2

# gpg: Signature made Di 28 Aug 2012 09:12:30 CEST
# gpg:                using DSA key 0x9FCF2CCD3F3E6426
# gpg: please do a --check-trustdb
# gpg: Good signature from "Guido Günther <agx@sigxcpu.org>"
# gpg:                 aka "Guido Guenther <agx@debian.org>"
# gpg:                 aka "Guido Guenther <agx@sigxcpu.org>"

+++ src/lxc/lxc_container.h
@@ -17,8 +17,8 @@
  * Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+ * License along with this library;  If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 
 #ifndef LXC_CONTAINER_H
@@ -51,7 +51,7 @@ int lxcContainerWaitForContinue(int control);
 
 int lxcContainerStart(virDomainDefPtr def,
                       virSecurityManagerPtr securityDriver,
-                      unsigned int nveths,
+                      size_t nveths,
                       char **veths,
                       int control,
                       int handshakefd,
